Consider shared networks when creating router in project
The current code did not consider shared networks when checking
if a network existed in a project. The simplest solution is to
to stop filtering for networks in a specific project and let
neutron return an error when we create the router.

Story: 2010886
Task: 48649
Change-Id: I8e2202b1c38b438b4075edd0555c4b9856c736be
(cherry picked from commit 89f616c110a898c627ec82d574f36122f4cc5528)