Remove MAC address requirement for node registration
This module currently does not allow for the creation of baremetal
nodes without including the NICs. We're proposing using this
module for enrolling nodes and then later running an out-of-band
Redfish inspection to get the NICs after enrolling the node in
Ironic.

The default option of the module will still be to move the node
to available state which will still require the NICs, however
an option `make_available_on_node_create` is provided to allow
leaving the node in enroll state.

Change-Id: I9b1636d778e7477191bfbe6094bd6bc70a4d384e